Add shopping cart functionality and design updates
A new shopping cart feature was implemented with functionality for adding and removing products, as well as updating product quantities. Design updates were also made to improve user experience, such as adding a shopping cart icon and enhancing cart page visualization. The cart button now scrolls to the top of the page when clicked.