Update TestimonialModal styles and fixes z-index
This commit modifies the Testimonial Modal by changing the class of the close button to "button-modal" and adjusting its styling in the CSS. It also fixes a potential overlay issue by changing the z-index value in Header.scss from 9999 to 99. Fixes have been applied to box-shadow and z-index properties to ensure proper display and function.