Implement Redux for cart functionality
This commit introduces Redux into the project to handle the cart feature. The cart slice was created for managing cart items, offering actions for adding, removing items, and increasing/decreasing item quantity. Usage of Redux is evident in both the Cart component, reflecting real quantity in the cart, and in the Product Card component, where adding products to cart was implemented. Also, UX improvements were made to the Cart Page, reflecting the real quantity and price of the selected items and enabling delete and quantity update actions.